requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
motor==3.6.0
pymongo==4.9.2
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
boto3==1.35.68
botocore==1.35.68
httpx==0.28.0
anthropic==0.39.0
python-dotenv==1.0.1
celery[sqs]==5.4.0
kombu==5.4.2
cryptography==43.0.3
python-multipart==0.0.17
python-magic==0.4.27
slowapi==0.1.9
limits==3.13.0
email-validator==2.2.0
google-auth==2.37.0
google-api-python-client==2.161.0
google-auth-httplib2==0.2.0
